requirements.txt
asgiref==3.7.2
Django==5.0.2
django-autoslug==1.9.9
django-cities-light==3.10.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
progressbar2==4.3.2
PyJWT==2.8.0
python-utils==3.8.2
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.9.0
Unidecode==1.3.8
